Fix: comparação de ícones arrumado tanto no backend quanto no frontend
