+++ FC_Backend_VS/src/features/categoria/categoria.controller.js
@@ -15,7 +15,10 @@ export class CategoriaController {
       const categorias = await this.service.findAll();
       return res.status(200).send({ sucesso: true, dados: categorias });
     } catch (err) {
-      return res.status(500).send({ sucesso: false, mensagem: "Erro interno" });
+      console.log("Erro ao listar categorias:", err);
+      return res
+        .status(500)
+        .send({ sucesso: false, mensagem: "Erro interno", stack: err.stack });
     }
   }
