+++ FC_Frontend/src/components/categoria_adm/Categoria_adm.module.css
@@ -4,7 +4,7 @@
     system-ui,
     -apple-system,
     sans-serif;
-  width: 280px;
+  width: 720px;
   display: flex;
   flex-direction: column;
   gap: 6px;
